+++ man-roxygen/as_is.R
@@ -1,1 +1,0 @@
-#' @param as_is logical. Whether to return the raw content from nettskjema. Default is FALSE.