+++ man-roxygen/asis.R
@@ -1,0 +1,1 @@
+#' @param asis logical. Whether to return the raw content from nettskjema. Default is FALSE.