+++ man-roxygen/data.R
@@ -1,1 +1,0 @@
-#' @param data tibble. Data retrieved from nettskjema