+++ man-roxygen/fields.R
@@ -1,1 +1,0 @@
-#' @param fields character vector of nettskjema fields