+++ tests/testthat.R
@@ -1,7 +1,5 @@
-library(testthat)
+# Load necessary libraries
 library(nettskjemar)
+library(testthat)
 
 test_check("nettskjemar")
-
-# cov <- covr::package_coverage()
-# covr::zero_coverage(cov)