+++ tests/testthat/helpers.R
@@ -1,5 +1,9 @@
-skip_if_no_auth <- function(token_name = "NETTSKJEMA_API_TOKEN") {
-  if (identical(Sys.getenv(token_name), "")) {
+form_id       <- 123823
+submission_id <- 27685292
+attachment_id <- 841793
+
+skip_if_no_auth <- function() {
+  if (identical(Sys.getenv("NETTSKJEMA_CLIENT_SECRET"), "")) {
     testthat::skip("No authentication available")
   }
 }